# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank project",
    "description": "第一步：通用数据接入与暂存。第二步：按表进行本体对齐、展示映射及生成独立图谱版本。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/alignment/config": {
      "get": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Config",
        "operationId": "config_api_v1_alignment_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alignment/runs": {
      "get": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Runs",
        "operationId": "runs_api_v1_alignment_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  },
                  "type": "array",
                  "title": "Response Runs Api V1 Alignment Runs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_v1_alignment_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alignment/runs/{run_id}": {
      "get": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Run",
        "operationId": "run_api_v1_alignment_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alignment/runs/{run_id}/concepts": {
      "get": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Concepts",
        "operationId": "concepts_api_v1_alignment_runs__run_id__concepts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConceptDetail"
                  },
                  "title": "Response Concepts Api V1 Alignment Runs  Run Id  Concepts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alignment/runs/{run_id}/graph": {
      "post": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Generate",
        "operationId": "generate_api_v1_alignment_runs__run_id__graph_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alignment/graph": {
      "get": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Graph",
        "operationId": "graph_api_v1_alignment_graph_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphPreview"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alignment/runs/{run_id}/template": {
      "post": {
        "tags": [
          "第二步 · 本体对齐与图谱生成"
        ],
        "summary": "Save Template",
        "operationId": "save_template_api_v1_alignment_runs__run_id__template_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphTemplate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/Selection"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Tables"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tables"
        ],
        "title": "AnalyzeRequest"
      },
...
      "ColumnMapping": {
        "properties": {
          "column": {
            "type": "string",
            "maxLength": 512,
            "title": "Column"
          },
          "role": {
            "type": "string",
            "enum": [
              "primary_key",
              "foreign_key",
              "attribute",
              "ignore"
            ],
            "title": "Role",
            "default": "attribute"
          },
          "semantic": {
            "type": "string",
            "maxLength": 80,
            "title": "Semantic",
            "default": ""
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "title": "Reason",
            "default": ""
          },
          "property_key": {
            "type": "string",
            "title": "Property Key"
          },
          "concept_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "column",
          "property_key"
        ],
        "title": "ColumnMapping"
      },
      "ConceptDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ConceptRef"
            },
            "type": "array",
            "title": "Parents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ConceptDetail"
      },
      "ConceptRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ConceptRef"
      },
...
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relation_id": {
            "type": "string",
            "title": "Relation Id"
          },
          "origin": {
            "type": "string",
            "enum": [
              "declared",
              "candidate",
              "confirmed"
            ],
            "title": "Origin"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "relation_id",
          "origin"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "table_id": {
            "type": "string",
            "title": "Table Id"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "source_row": {
            "type": "integer",
            "title": "Source Row"
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "fields": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "table_id",
          "table_name",
          "source_row",
          "concept_id",
          "concept_name",
          "fields"
        ],
        "title": "GraphNode"
      },
      "GraphPreview": {
        "properties": {
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "title": "GraphPreview"
      },
      "GraphSummary": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count"
          },
          "edge_count": {
            "type": "integer",
            "title": "Edge Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "version",
          "run_id",
          "revision",
          "node_count",
          "edge_count",
          "created_at"
        ],
        "title": "GraphSummary"
      },
      "GraphTemplate": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TemplateNode"
            },
            "type": "array",
            "maxItems": 120,
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/TemplateEdge"
            },
            "type": "array",
            "maxItems": 240,
            "title": "Edges"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GraphTemplate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MappingResult": {
        "properties": {
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphTemplate"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "snapshot_sha256": {
            "type": "string",
            "title": "Snapshot Sha256"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableMapping"
            },
            "type": "array",
            "title": "Tables"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationMapping"
            },
            "type": "array",
            "title": "Relations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "snapshot_sha256",
          "tables",
          "relations"
        ],
        "title": "MappingResult"
      },
      "MatchStep": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "meaning",
              "recall",
              "selection",
              "validation"
            ],
            "title": "Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "skipped"
            ],
            "title": "Status",
            "default": "pending"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "MatchStep"
      },
      "MatchTrace": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "model",
              "retrieve"
            ],
            "title": "Method",
            "default": "model"
          },
          "retrievals": {
            "items": {
              "$ref": "#/components/schemas/RetrievalTrace"
            },
            "type": "array",
            "title": "Retrievals"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/MatchStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "sample_rows": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sample Rows"
          },
          "meaning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TableMeaning"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ConceptDetail"
            },
            "type": "array",
            "title": "Candidates"
          },
          "attribute_candidates": {
            "items": {
              "$ref": "#/components/schemas/ConceptDetail"
            },
            "type": "array",
            "title": "Attribute Candidates"
          },
          "selected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConceptDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "selection_reason": {
            "type": "string",
            "title": "Selection Reason",
            "default": ""
          },
          "selection_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Confidence"
          },
          "verification": {
            "type": "string",
            "title": "Verification",
            "default": "none"
          },
          "selection_attempts": {
            "type": "integer",
            "title": "Selection Attempts",
            "default": 0
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold"
          }
        },
        "type": "object",
        "required": [
          "confidence_threshold"
        ],
        "title": "MatchTrace"
      },
...
      "RelationMapping": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_table_id": {
            "type": "string",
            "title": "Source Table Id"
          },
          "target_table_id": {
            "type": "string",
            "title": "Target Table Id"
          },
          "source_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Columns"
          },
          "target_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Columns"
          },
          "origin": {
            "type": "string",
            "enum": [
              "declared",
              "candidate"
            ],
            "title": "Origin"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "skipped"
            ],
            "title": "Status"
          },
          "matched_rows": {
            "type": "integer",
            "title": "Matched Rows",
            "default": 0
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_table_id",
          "target_table_id",
          "source_columns",
          "target_columns",
          "origin",
          "reason",
          "status"
        ],
        "title": "RelationMapping"
      },
...
      "Run": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "enum": [
              "analyzing",
              "ready",
              "failed"
            ],
            "title": "Status",
            "default": "analyzing"
          },
          "progress": {
            "type": "string",
            "title": "Progress",
            "default": "等待分析"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MappingResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_status": {
            "type": "string",
            "enum": [
              "none",
              "building",
              "ready",
              "failed"
            ],
            "title": "Graph Status",
            "default": "none"
          },
          "graph_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Error"
          },
          "graph_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Version"
          },
          "based_on_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Run Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "Run"
      },
...
      "Selection": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "table_id": {
            "type": "string",
            "format": "uuid",
            "title": "Table Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "table_id"
        ],
        "title": "Selection"
      },
...
      "TableMapping": {
        "properties": {
          "table_id": {
            "type": "string",
            "title": "Table Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "concept_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Name"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "mapped",
              "review",
              "unmatched",
              "failed"
            ],
            "title": "Status",
            "default": "unmatched"
          },
          "verification": {
            "type": "string",
            "enum": [
              "local_catalog",
              "verified",
              "mismatch",
              "unavailable",
              "none",
              "manual"
            ],
            "title": "Verification",
            "default": "none"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnMapping"
            },
            "type": "array",
            "title": "Columns"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "structure_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Structure Notes"
          },
          "manual_edits": {
            "items": {
              "$ref": "#/components/schemas/MappingEdit"
            },
            "type": "array",
            "title": "Manual Edits"
          }
        },
        "type": "object",
        "required": [
          "table_id",
          "batch_id",
          "table_name",
          "row_count"
        ],
        "title": "TableMapping"
      },
...
      "TemplateEdge": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source"
          },
          "target": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Target"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "same_row",
              "join"
            ],
            "title": "Mode",
            "default": "same_row"
          },
          "source_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Source Columns"
          },
          "target_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Target Columns"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "name",
          "reason"
        ],
        "title": "TemplateEdge"
      },
      "TemplateNode": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Id"
          },
          "table_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Table Id"
          },
          "concept_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Concept Id"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name",
            "default": ""
          },
          "identity_scope": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Identity Scope"
          },
          "key_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Key Columns"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/TemplateProperty"
            },
            "type": "array",
            "maxItems": 2048,
            "title": "Properties"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "table_id",
          "concept_id",
          "identity_scope"
        ],
        "title": "TemplateNode"
      },
      "TemplateProperty": {
        "properties": {
          "column": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Column"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "column",
          "name"
        ],
        "title": "TemplateProperty"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}